chore(natpmp): apply review comments round 2
- rename retryAfterReadError to handleReadError and retryErr to fatalErr
- make reserveClosedPort compile on all platforms by splitting it into a
  linux/darwin implementation and a stub which panics otherwise
- clarify the reserveClosedPort comment about why a TCP socket is used
