+++ internal/natpmp/helpers_test.go
@@ -6,55 +6,12 @@ import (
 	"testing"
 	"time"
 
-	"github.com/qdm12/gluetun/internal/pmtud/constants"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"golang.org/x/sys/unix"
 )
 
 // enough for slow machines for local UDP server.
 const initialConnectionDuration = 3 * time.Second
-
-// reserveClosedPort returns a UDP port which is reserved by a TCP socket
-// which is bound but not listening, so that no UDP process is listening on
-// it and the OS cannot assign the port to another process. The OS sends
-// back an ICMP port unreachable message for datagrams sent to that port,
-// which surfaces on the sender as a connection refused error on read.
-func reserveClosedPort(t *testing.T) (port uint16) {
-	t.Helper()
-
-	fd, err := unix.Socket(constants.AF_INET, constants.SOCK_STREAM, constants.IPPROTO_TCP)
-	require.NoError(t, err)
-	t.Cleanup(func() {
-		err := unix.Close(fd)
-		assert.NoError(t, err)
-	})
-
-	addr := &unix.SockaddrInet4{
-		Port: 0,
-		Addr: [4]byte{127, 0, 0, 1},
-	}
-
-	err = unix.Bind(fd, addr)
-	if err != nil {
-		_ = unix.Close(fd)
-		t.Fatal(err)
-	}
-
-	sockAddr, err := unix.Getsockname(fd)
-	if err != nil {
-		_ = unix.Close(fd)
-		t.Fatal(err)
-	}
-
-	sockAddr4, ok := sockAddr.(*unix.SockaddrInet4)
-	if !ok {
-		_ = unix.Close(fd)
-		t.Fatal("not an IPv4 address")
-	}
-
-	return uint16(sockAddr4.Port) //nolint:gosec
-}
 
 type udpExchange struct {
 	request  []byte
